requirements.txt
bidict==0.23.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
packaging==24.2
pillow==11.0.0
pymongo==4.10.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
qrcode==8.0
redis==5.2.0
requests==2.32.3
simple-websocket==1.1.0
spotipy==2.24.0
urllib3==2.2.3
Werkzeug==3.0.6
wsproto==1.2.0